Refactor MovementResolver
  ref #7
