+++ lib/variant/standard/movement-resolver.js
@@ -1,4 +1,3 @@
-const DirectedGraph = require('../../util/directed-graph')
 const Unit = require('../../data/unit')
 const State = require('../../data/state')
 const Board = require('../../data/board')
@@ -6,6 +5,7 @@ const RuleHelper = require('../../rule/rule-helper')
 const StandardRuleUtil = require('./standard-rule-util')
 
 const MovementOrderGroup = require('./movement-order-group')
+const OrderDependencyGraph = require('./order-dependency-graph')
 
 const AbleToConvoy = 'AbleToConvoy'
 
@@ -91,65 +91,9 @@ module.exports = class MovementResolver {
     })
 
     // 3. Generate the dependency graph
-    const nodeMap = new Map()
-    const edges = []
+    let graph = new OrderDependencyGraph(this.rule, orders, orderResult)
 
-    function addNode (province) {
-      if (!nodeMap.has(province)) {
-        nodeMap.set(province, nodeMap.size)
-      }
-    }
-
-    orders.forEach(order => {
-      switch (order.type) {
-        case 'Hold':
-          addNode(order.unit.location.province)
-          break
-        case 'Move':
-          addNode(order.unit.location.province)
-          addNode(order.destination.province)
-          edges.push(
-            [nodeMap.get(order.destination.province), nodeMap.get(order.unit.location.province)]
-          )
-
-          break
-        case 'Support':
-          const destination = StandardRuleUtil.getDestinationOfSupportOrder(order)
-          addNode(order.unit.location.province)
-          addNode(destination.province)
-          if (orderResult.get(order) !== $.NoCorrespondingOrder) {
-            edges.push(
-              [nodeMap.get(order.unit.location.province), nodeMap.get(destination.province)]
-            )
-          }
-          break
-        case 'Convoy':
-          addNode(order.unit.location.province)
-          addNode(order.target.destination.province)
-          if (orderResult.get(order) !== $.NoCorrespondingOrder) {
-            edges.push(
-              [
-                nodeMap.get(order.unit.location.province),
-                nodeMap.get(order.target.destination.province)
-              ]
-            )
-          }
-          break
-      }
-    })
-    const nodes = new Map([...nodeMap].map(elem => [elem[1], [elem[0]]]))
-    let graph = new DirectedGraph(nodes, edges)
-
-    // 4. Find cycles and merge nodes
-    while (true) {
-      const cycle = graph.getCycle()
-      if (!cycle) break
-      if (cycle.length <= 1) break
-
-      graph = graph.mergeNodes(cycle)
-    }
-
-    // 5. Resolve each province
+    // 4. Resolve each province
     while (graph.nodes.size !== 0) {
       const node = [...graph.nodes].find(node => graph.edges.every(edge => edge[1] !== node[0]))
       if (!node) return `Internal Error: ${orders}`
@@ -344,8 +288,6 @@ module.exports = class MovementResolver {
 
     // Geneate a new board
     const { state, units, occupation } = board
-    let nextState = null
-    let nextOccupation = occupation
 
     const nextUnits = new Map()
     const nextUnitsStatuses = new Map()
@@ -384,20 +326,13 @@ module.exports = class MovementResolver {
       }
     })
 
-    const existsDislodgedUnits = [...orderResult].some(elem => elem[1] === $.Dislodged)
-    if (state.season === $.Autumn) {
-      nextOccupation = StandardRuleUtil.updateOccupation(nextUnits, occupation)
-      nextState =
-        (existsDislodgedUnits)
-          ? new State(state.year, state.season, $.Retreat)
-          : (StandardRuleUtil.isBuildPhaseRequired(nextUnits, nextOccupation))
-            ? new State(state.year, state.season, $.Build)
-            : new State(state.year, $.Autumn, $.Movement)
-    } else {
-      nextState = (existsDislodgedUnits)
-        ? new State(state.year, state.season, $.Retreat)
-        : new State(state.year, $.Autumn, $.Movement)
-    }
+    const nextOccupation =
+      (state.season === $.Autumn)
+        ? StandardRuleUtil.updateOccupation(nextUnits, occupation)
+        : occupation
+
+    const nextState =
+      StandardRuleUtil.getNextState(this.rule, state, nextUnits, nextOccupation, nextUnitsStatuses)
 
     const nextBoard =
       new Board(nextState, nextUnits, nextOccupation, nextUnitsStatuses, nextProvinceSatuses)
