+++ lib/variant/standard/standard-rule-util.js
@@ -1,3 +1,6 @@
+const RuleHelper = require('../../rule/rule-helper')
+const State = require('../../data/state')
+
 module.exports = class StandardRuleUtil {
   static getDestinationOfSupportOrder (support) {
     return (support.target.type === 'Hold')
@@ -56,5 +59,24 @@ module.exports = class StandardRuleUtil {
     })
 
     return (x) ? x[0] : x
+  }
+  static getNextState (rule, state, units, occupation, unitStatuses) {
+    const $ = new RuleHelper(rule)
+
+    if (state.season === $.Autumn) {
+      if ([...unitStatuses].some(elem => elem[1] === $.Dislodged)) {
+        return new State(state.year, state.season, $.Retreat)
+      } else if (StandardRuleUtil.isBuildPhaseRequired(units, occupation)) {
+        return new State(state.year, state.season, $.Build)
+      } else {
+        return new State(state.year + 1, $.Spring, $.Movement)
+      }
+    } else {
+      if ([...unitStatuses].some(elem => elem[1] === $.Dislodged)) {
+        return new State(state.year, state.season, $.Retreat)
+      } else {
+        return new State(state.year, $.Autumn, $.Movement)
+      }
+    }
   }
 }
